4.1. Разбор первой программы. [#159033 #99687]

+++ src/main/java/ru/job4j/Multiple.java
@@ -2,8 +2,8 @@ package ru.job4j;
 
 public class Multiple {
     public static  void main (String[] args) {
-        int x = 1;
-        for (int y = 1; y<=10; y++)
+        byte x = 1;
+        for (byte y = 1; y<=10; y++)
 
             System.out.println(x + " * " + y + " = " + (x * y) + "\t");
         }
